ultimo cambio compras

+++ ms_compras/src/main/java/com/example/ms_compras/service/impl/ComprasServiceImpl.java
@@ -21,7 +21,7 @@ public class ComprasServiceImpl implements ComprasService {
 
     @Override
     public Optional<Compras> buscarPorId(Integer id) {
-        return categoriaRepository.findById(id);
+        return categoriaRepository.findById(id.longValue());
     }
 
     @Override
@@ -36,6 +36,6 @@ public class ComprasServiceImpl implements ComprasService {
 
     @Override
     public void borrarPorId(Integer id) {
-        categoriaRepository.deleteById(id);
+        categoriaRepository.deleteById(id.longValue());
     }
 }
